Create interfaces for Audio and Input on Lwjgl3 backend
