+++ backends/gdx-backend-lwjgl3/src/com/badlogic/gdx/backends/lwjgl3/Lwjgl3Application.java
@@ -21,6 +21,8 @@ import java.io.PrintStream;
 import java.nio.IntBuffer;
 
 import com.badlogic.gdx.ApplicationLogger;
+import com.badlogic.gdx.backends.lwjgl3.audio.Lwjgl3Audio;
+import com.badlogic.gdx.backends.lwjgl3.audio.OpenALLwjgl3Audio;
 import com.badlogic.gdx.graphics.glutils.GLVersion;
 
 import org.lwjgl.BufferUtils;
@@ -47,7 +49,6 @@ import com.badlogic.gdx.Input;
 import com.badlogic.gdx.LifecycleListener;
 import com.badlogic.gdx.Net;
 import com.badlogic.gdx.Preferences;
-import com.badlogic.gdx.backends.lwjgl3.audio.OpenALAudio;
 import com.badlogic.gdx.backends.lwjgl3.audio.mock.MockAudio;
 import com.badlogic.gdx.utils.Array;
 import com.badlogic.gdx.utils.Clipboard;
@@ -55,11 +56,11 @@ import com.badlogic.gdx.utils.GdxRuntimeException;
 import com.badlogic.gdx.utils.ObjectMap;
 import com.badlogic.gdx.utils.SharedLibraryLoader;
 
-public class Lwjgl3Application implements Application {
+public class Lwjgl3Application implements Lwjgl3ApplicationBase {
 	private final Lwjgl3ApplicationConfiguration config;
 	final Array<Lwjgl3Window> windows = new Array<Lwjgl3Window>();
 	private volatile Lwjgl3Window currentWindow;
-	private Audio audio;
+	private Lwjgl3Audio audio;
 	private final Files files;
 	private final Net net;
 	private final ObjectMap<String, Preferences> preferences = new ObjectMap<String, Preferences>();
@@ -94,15 +95,15 @@ public class Lwjgl3Application implements Application {
 		Gdx.app = this;
 		if (!config.disableAudio) {
 			try {
-				this.audio = Gdx.audio = new OpenALAudio(config.audioDeviceSimultaneousSources,
-						config.audioDeviceBufferCount, config.audioDeviceBufferSize);
+				this.audio = createAudio(config);
 			} catch (Throwable t) {
 				log("Lwjgl3Application", "Couldn't initialize audio, disabling audio", t);
-				this.audio = Gdx.audio = new MockAudio();
+				this.audio = new MockAudio();
 			}
 		} else {
-			this.audio = Gdx.audio = new MockAudio();
+			this.audio = new MockAudio();
 		}
+		Gdx.audio = audio;
 		this.files = Gdx.files = new Lwjgl3Files();
 		this.net = Gdx.net = new Lwjgl3Net(config);
 		this.clipboard = new Lwjgl3Clipboard();
@@ -126,9 +127,7 @@ public class Lwjgl3Application implements Application {
 		Array<Lwjgl3Window> closedWindows = new Array<Lwjgl3Window>();
 		while (running && windows.size > 0) {
 			// FIXME put it on a separate thread
-			if (audio instanceof OpenALAudio) {
-				((OpenALAudio) audio).update();
-			}
+			audio.update();
 
 			boolean haveWindowsRendered = false;
 			closedWindows.clear();
@@ -207,9 +206,7 @@ public class Lwjgl3Application implements Application {
 	
 	private void cleanup() {
 		Lwjgl3Cursor.disposeSystemCursors();
-		if (audio instanceof OpenALAudio) {
-			((OpenALAudio) audio).dispose();
-		}
+		audio.dispose();
 		errorCallback.free();
 		errorCallback = null;
 		if (glDebugCallback != null) {
@@ -361,7 +358,18 @@ public class Lwjgl3Application implements Application {
 			lifecycleListeners.removeValue(listener, true);
 		}
 	}
-	
+
+	@Override
+	public Lwjgl3Audio createAudio (Lwjgl3ApplicationConfiguration config) {
+		return new OpenALLwjgl3Audio(config.audioDeviceSimultaneousSources,
+			config.audioDeviceBufferCount, config.audioDeviceBufferSize);
+	}
+
+	@Override
+	public Lwjgl3Input createInput (Lwjgl3Window window) {
+		return new DefaultLwjgl3Input(window);
+	}
+
 	/**
 	 * Creates a new {@link Lwjgl3Window} using the provided listener and {@link Lwjgl3WindowConfiguration}.
 	 *
@@ -376,7 +384,7 @@ public class Lwjgl3Application implements Application {
 
 	private Lwjgl3Window createWindow (final Lwjgl3ApplicationConfiguration config, ApplicationListener listener,
 		final long sharedContext) {
-		final Lwjgl3Window window = new Lwjgl3Window(listener, config);
+		final Lwjgl3Window window = new Lwjgl3Window(listener, config, this);
 		if (sharedContext == 0) {
 			// the main window is created immediately
 			createWindow(window, config, sharedContext);
