+++ backends/gdx-backend-lwjgl3/src/com/badlogic/gdx/backends/lwjgl3/Lwjgl3Window.java
@@ -40,6 +40,7 @@ import com.badlogic.gdx.utils.SharedLibraryLoader;
 public class Lwjgl3Window implements Disposable {
 	private long windowHandle;
 	final ApplicationListener listener;
+	final Lwjgl3ApplicationBase application;
 	private boolean listenerInitialized = false;
 	Lwjgl3WindowListener windowListener;
 	private Lwjgl3Graphics graphics;
@@ -153,17 +154,18 @@ public class Lwjgl3Window implements Disposable {
 		}
 	};
 
-	Lwjgl3Window(ApplicationListener listener, Lwjgl3ApplicationConfiguration config) {
+	Lwjgl3Window(ApplicationListener listener, Lwjgl3ApplicationConfiguration config, Lwjgl3ApplicationBase application) {
 		this.listener = listener;
 		this.windowListener = config.windowListener;
 		this.config = config;
+		this.application = application;
 		this.tmpBuffer = BufferUtils.createIntBuffer(1);
 		this.tmpBuffer2 = BufferUtils.createIntBuffer(1);
 	}
 
 	void create(long windowHandle) {
 		this.windowHandle = windowHandle;
-		this.input = new Lwjgl3Input(this);
+		this.input = application.createInput(this);
 		this.graphics = new Lwjgl3Graphics(this);
 
 		GLFW.glfwSetWindowFocusCallback(windowHandle, focusCallback);
