+++ backends/gdx-backend-lwjgl3/src/com/badlogic/gdx/backends/lwjgl3/audio/Lwjgl3Audio.java
@@ -1,0 +1,9 @@
+package com.badlogic.gdx.backends.lwjgl3.audio;
+
+import com.badlogic.gdx.Audio;
+import com.badlogic.gdx.utils.Disposable;
+
+public interface Lwjgl3Audio extends Audio, Disposable {
+
+	void update ();
+}
