+++ backends/gdx-backend-lwjgl3/src/com/badlogic/gdx/backends/lwjgl3/audio/OpenALSound.java
@@ -26,10 +26,10 @@ import static org.lwjgl.openal.AL10.*;
 /** @author Nathan Sweet */
 public class OpenALSound implements Sound {
 	private int bufferID = -1;
-	private final OpenALAudio audio;
+	private final OpenALLwjgl3Audio audio;
 	private float duration;
 
-	public OpenALSound (OpenALAudio audio) {
+	public OpenALSound (OpenALLwjgl3Audio audio) {
 		this.audio = audio;
 	}
